backend/requirements.txt
aiohttp>=3.12.15
chromadb>=1.1.0
cryptography>=41.0.0
fastapi>=0.116.2
gitpython>=3.1.45
jinja2>=3.1.6
langchain>=0.3.27
langchain-community>=0.3.29
langchain-openai>=0.3.33
pocketflow>=0.0.3
pymysql>=1.1.2
python-dotenv>=1.1.1
python-multipart>=0.0.20
requests>=2.32.5
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.23
alembic>=1.13.1
celery>=5.3.0
redis>=5.0.0
